src/main/java/dev/andrenascimento/case_pratico_java_springboot/dtos/ProdutoResponse.java
package dev.andrenascimento.case_pratico_java_springboot.dtos;

// Para enviar dados de produtos como resposta da API.
public class ProdutoResponse {
    private Long id;
    private String nome;
    private Double preco;
    private String descricao;
    private Integer quantidadeEmEstoque;

    
    public ProdutoResponse() {
    }

    public ProdutoResponse(Long id, String nome, Double preco, String descricao, Integer quantidadeEmEstoque) {
        this.id = id;
        this.nome = nome;
        this.preco = preco;
        this.descricao = descricao;
        this.quantidadeEmEstoque = quantidadeEmEstoque;
    }

    // Getters e Setters
    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    public Double getPreco() {
        return preco;
    }

    public void setPreco(Double preco) {
        this.preco = preco;
    }

    public String getDescricao() {
        return descricao;
    }

    public void setDescricao(String descricao) {
        this.descricao = descricao;
    }

    public Integer getQuantidadeEmEstoque() {
        return quantidadeEmEstoque;
    }

    public void setQuantidadeEmEstoque(Integer quantidadeEmEstoque) {
        this.quantidadeEmEstoque = quantidadeEmEstoque;
    }
}